seva/mvsplat/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.9.2
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
addict=2.4.0=pypi_0
aiofiles=23.2.1=pypi_0
aiohappyeyeballs=2.6.1=pypi_0
aiohttp=3.11.18=pypi_0
aiosignal=1.3.2=pypi_0
annotated-types=0.7.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=4.9.0=pypi_0
asttokens=3.0.0=pypi_0
async-timeout=5.0.1=pypi_0
attrs=25.3.0=pypi_0
beartype=0.20.2=pypi_0
beautifulsoup4=4.13.4=pypi_0
black=25.1.0=pypi_0
blinker=1.9.0=pypi_0
browser-cookie3=0.20.1=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.2.25=h06a4308_0
certifi=2025.1.31=pypi_0
charset-normalizer=3.4.1=pypi_0
click=8.1.8=pypi_0
colorama=0.4.6=pypi_0
colorlog=6.9.0=pypi_0
colorspacious=1.1.2=pypi_0
comm=0.2.2=pypi_0
configargparse=1.7=pypi_0
contourpy=1.3.2=pypi_0
cycler=0.12.1=pypi_0
dacite=1.9.2=pypi_0
dash=3.0.4=pypi_0
decorator=5.2.1=pypi_0
diff-gaussian-rasterization=0.0.0=pypi_0
diffusers=0.32.2=pypi_0
docker-pycreds=0.4.0=pypi_0
docstring-parser=0.16=pypi_0
e3nn=0.5.6=pypi_0
einops=0.8.1=pypi_0
embreex=2.17.7.post6=pypi_0
exceptiongroup=1.2.2=pypi_0
executing=2.2.0=pypi_0
fastapi=0.115.12=pypi_0
fastjsonschema=2.21.1=pypi_0
ffmpy=0.5.0=pypi_0
filelock=3.13.1=pypi_0
fire=0.7.0=pypi_0
flask=3.0.3=pypi_0
fonttools=4.57.0=pypi_0
frozenlist=1.6.0=pypi_0
fsspec=2024.6.1=pypi_0
ftfy=6.3.1=pypi_0
gdown=5.2.0=pypi_0
gitdb=4.0.12=pypi_0
gitpython=3.1.44=pypi_0
gradio=5.17.0=pypi_0
gradio-client=1.7.1=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.7=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=0.29.3=pypi_0
hydra-core=1.3.2=pypi_0
idna=3.10=pypi_0
imageio=2.37.0=pypi_0
imageio-ffmpeg=0.6.0=pypi_0
importlib-metadata=8.6.1=pypi_0
ipython=8.36.0=pypi_0
ipywidgets=8.1.7=pypi_0
itsdangerous=2.2.0=pypi_0
jaxtyping=0.3.2=pypi_0
jedi=0.19.2=pypi_0
jeepney=0.9.0=pypi_0
jinja2=3.1.4=pypi_0
joblib=1.5.0=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
jupyter-core=5.7.2=pypi_0
jupyterlab-widgets=3.0.15=pypi_0
kiwisolver=1.4.8=pypi_0
kornia=0.8.0=pypi_0
kornia-rs=0.1.8=pypi_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
lightning-utilities=0.14.3=pypi_0
lpips=0.1.4=pypi_0
lxml=5.3.1=pypi_0
lz4=4.4.4=pypi_0
manifold3d=3.0.1=pypi_0
mapbox-earcut=1.0.3=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.10.1=pypi_0
matplotlib-inline=0.1.7=pypi_0
mdurl=0.1.2=pypi_0
moviepy=2.1.2=pypi_0
mpmath=1.3.0=pypi_0
msgspec=0.19.0=pypi_0
multidict=6.4.3=pypi_0
mypy-extensions=1.1.0=pypi_0
narwhals=1.38.2=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pypi_0
networkx=3.3=pypi_0
ninja=1.11.1.4=pypi_0
nodeenv=1.9.1=pypi_0
numpy=2.2.5=pypi_0
nvidia-cublas-cu11=11.11.3.6=pypi_0
nvidia-cuda-cupti-cu11=11.8.87=pypi_0
nvidia-cuda-nvrtc-cu11=11.8.89=pypi_0
nvidia-cuda-runtime-cu11=11.8.89=pypi_0
nvidia-cudnn-cu11=9.1.0.70=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-curand-cu11=10.3.0.86=pypi_0
nvidia-cusolver-cu11=11.4.1.48=pypi_0
nvidia-cusparse-cu11=11.7.5.86=pypi_0
nvidia-nccl-cu11=2.21.5=pypi_0
nvidia-nvtx-cu11=11.8.86=pypi_0
omegaconf=2.3.0=pypi_0
open-clip-torch=2.31.0=pypi_0
open3d=0.19.0=pypi_0
opencv-python-headless=4.11.0.86=pypi_0
openssl=3.0.16=h5eee18b_0
opt-einsum=3.4.0=pypi_0
opt-einsum-fx=0.1.4=pypi_0
orjson=3.10.16=pypi_0
outcome=1.3.0.post0=pypi_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
parso=0.8.4=pypi_0
pathspec=0.12.1=pypi_0
pexpect=4.9.0=pypi_0
pillow=10.4.0=pypi_0
pip=25.0=py310h06a4308_0
platformdirs=4.3.8=pypi_0
plotly=6.0.1=pypi_0
plyfile=1.1=pypi_0
proglog=0.1.11=pypi_0
prompt-toolkit=3.0.51=pypi_0
propcache=0.3.1=pypi_0
protobuf=6.30.2=pypi_0
psutil=7.0.0=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.3=pypi_0
pycollada=0.9=pypi_0
pycryptodomex=3.22.0=pypi_0
pydantic=2.10.6=pypi_0
pydantic-core=2.27.2=pypi_0
pydub=0.25.1=pypi_0
pygments=2.19.1=pypi_0
pyparsing=3.2.3=pypi_0
pyquaternion=0.9.9=pypi_0
pysocks=1.7.1=pypi_0
python=3.10.16=he870216_1
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.1.0=pypi_0
python-multipart=0.0.20=pypi_0
pytorch-lightning=2.5.1.post0=pypi_0
pytz=2025.1=pypi_0
pyyaml=6.0.2=pypi_0
readline=8.2=h5eee18b_0
referencing=0.36.2=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
retrying=1.3.4=pypi_0
rich=13.9.4=pypi_0
roma=1.5.2.1=pypi_0
rpds-py=0.23.1=pypi_0
rtree=1.4.0=pypi_0
ruff=0.11.2=pypi_0
safehttpx=0.1.6=pypi_0
safetensors=0.5.3=pypi_0
scikit-image=0.25.2=pypi_0
scikit-learn=1.6.1=pypi_0
scipy=1.15.2=pypi_0
selenium=4.32.0=pypi_0
semantic-version=2.10.0=pypi_0
sentry-sdk=2.27.0=pypi_0
setproctitle=1.3.6=pypi_0
setuptools=75.8.0=py310h06a4308_0
seva=0.0.0=pypi_0
shapely=2.0.7=pypi_0
shellingham=1.5.4=pypi_0
shtab=1.7.1=pypi_0
six=1.17.0=pypi_0
sk-video=1.1.10=pypi_0
smmap=5.0.2=pypi_0
sniffio=1.3.1=pypi_0
sortedcontainers=2.4.0=pypi_0
soupsieve=2.7=pypi_0
splines=0.3.2=pypi_0
sqlite=3.45.3=h5eee18b_0
stack-data=0.6.3=pypi_0
starlette=0.46.1=pypi_0
svg-path=6.3=pypi_0
svg-py=1.6.0=pypi_0
sympy=1.13.1=pypi_0
tabulate=0.9.0=pypi_0
termcolor=2.5.0=pypi_0
threadpoolctl=3.6.0=pypi_0
tifffile=2025.3.13=pypi_0
timm=1.0.15=pypi_0
tk=8.6.14=h39e8969_0
tomli=2.2.1=pypi_0
tomlkit=0.13.2=pypi_0
torch=2.6.0+cu118=pypi_0
torchaudio=2.6.0+cu118=pypi_0
torchmetrics=1.7.1=pypi_0
torchvision=0.21.0+cu118=pypi_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pypi_0
trimesh=4.6.5=pypi_0
trio=0.30.0=pypi_0
trio-websocket=0.12.2=pypi_0
triton=3.2.0=pypi_0
typeguard=4.4.2=pypi_0
typer=0.15.2=pypi_0
typing-extensions=4.12.2=pypi_0
tyro=0.9.17=pypi_0
tzdata=2025.2=pypi_0
urllib3=2.3.0=pypi_0
uvicorn=0.34.0=pypi_0
vhacdx=0.0.8.post2=pypi_0
viser=0.2.23=pypi_0
wadler-lindig=0.1.5=pypi_0
wandb=0.19.11=pypi_0
wcwidth=0.2.13=pypi_0
websocket-client=1.8.0=pypi_0
websockets=14.2=pypi_0
werkzeug=3.0.6=pypi_0
wheel=0.45.1=py310h06a4308_0
widgetsnbextension=4.0.14=pypi_0
wsproto=1.2.0=pypi_0
xatlas=0.0.9=pypi_0
xxhash=3.5.0=pypi_0
xz=5.6.4=h5eee18b_1
yarl=1.20.0=pypi_0
yourdfpy=0.0.57=pypi_0
youtube-dl=2021.12.17=pypi_0
yt-dlp=2025.4.30=pypi_0
zipp=3.21.0=pypi_0
zlib=1.2.13=h5eee18b_1
